improvments

+++ src/main/java/com/printshopmanagement/application/domain/Employee.java
@@ -5,8 +5,10 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
 
 @Getter
 @Setter
@@ -17,7 +19,7 @@ public class Employee{
 
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
 
     @Column
     private Long employeePersonalNumber;
